requirements.txt
requests>=2.28.0
pydantic>=2.11.9
python-dotenv==1.2.1
